backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
sqlalchemy==2.0.21
alembic==1.12.0
pydantic-settings==2.0.3
python-multipart==0.0.6
pillow==10.0.1
opencv-python-headless==4.8.1.78
numpy==1.24.4
torch==2.0.1
open-clip-torch==2.20.0
scikit-learn==1.3.1
python-dotenv==1.0.0
loguru==0.7.1
